Move the InterruptedInListenerTest to ListenersTest
