+++ testng-core/src/test/resources/testng.xml
@@ -994,10 +994,4 @@
       <class name="org.testng.BasicTest"/>
     </classes>
   </test>
-
-  <test name="GITHUB-2685">
-    <classes>
-      <class name="test.listeners.issue2685.InterruptedInListenerTest"/>
-    </classes>
-  </test>
 </suite>
